fixed comparing

+++ modules/kiselev_i_billinear_interpolation/src/Color.cpp
@@ -3,16 +3,16 @@
 #include "include/Color.h"
 
 Color::Color(int r, int g, int b) {
-    if (r > max) R = max;
-    else if (r < min) R = min;
+    if (r > (int)max) R = max;
+    else if (r < (int)min) R = min;
     else
             R = r;
-    if (g > max) G = max;
-    else if (g < min) G = min;
+    if (g > (int)max) G = max;
+    else if (g < (int)min) G = min;
     else
             G = g;
-    if (b > max) B = max;
-    else if (b < min) B = min;
+    if (b > (int)max) B = max;
+    else if (b < (int)min) B = min;
     else
             B = b;
 }
@@ -36,22 +36,22 @@ char Color::getB() const {
 }
 
 void Color::setR(int r) {
-    if (r > max) R = max;
-    else if (r < min) R = min;
+    if (r > (int)max) R = max;
+    else if (r < (int)min) R = min;
     else
             R = r;
 }
 
 void Color::setG(int g) {
-    if (g > max) G = max;
-    else if (g < min) G = min;
+    if (g > (int)max) G = max;
+    else if (g < (int)min) G = min;
     else
             G = g;
 }
 
 void Color::setB(int b) {
-    if (b > max) B = max;
-    else if (b < min) B = min;
+    if (b > (int)max) B = max;
+    else if (b < (int)min) B = min;
     else
             B = b;
 }
